Use tmp buffer to ensure snippet table is loaded only if required
Because yasnippets are JIT loaded, we were earlier creating a tmp
buffer and activating the major mode in it to ensure loading of the
snippet table for the major mode. This was however done
everytime. This commit optimizes the code by resorting to the
workaround only if the snippet table has not been already loaded.