Use 'cl-lib' instead of 'dash' for sequence abstractions
Reason for this change: 'cl-lib' is built into emacs and 'yasnippet'
already depends on it. This means axy doesn't have any additional
dependency besides 'yasnippet'. It also means that if the maintainers
of 'yasnippet' feel the code is worthy of merging in 'yasnippet'
itself, that can also be done without introducing any additional
dependency.